+++ lib/src/scope.dart
@@ -8,12 +8,35 @@ import 'package:meta/meta.dart';
 import 'exceptions.dart';
 
 part 'scope_key.dart';
-part 'injector.dart';
-part 'single_injector.dart';
+part '__injector.dart';
+part '__single_injector.dart';
 
-typedef Factory = dynamic Function();
+typedef _Factory = dynamic Function();
 
+/// Creates a Scope providing dependency injection to your call stack.
+///
+/// Scopes may be nested with the nearest Scope overriding parent scopes.
 class Scope {
+  /// Create a [Scope] that allows you to inject values.
+  ///
+  /// Any methods directly or indirectly called from the
+  /// [Scope]s [run] method have access to those injected values.
+  ///
+  /// The [debugName] is useful when debugging allowing yo to
+  /// provide each [Scope] with a unique name.
+  ///
+  /// ```dart
+  /// final ageKey = ScopeKey<int>();
+  /// final daysOldKey = ScopeKey<int>();
+  /// final countKey = ScopeKey<int>();
+  /// Scope()
+  ///   ..value<int>(ageKey, 18)
+  ///   ..single<int>(daysOldKey, () => calculateDaysOld(use(ageKey)))
+  ///   ..sequence<int>(countKey, () => count++)
+  ///   ..run(() {
+  ///       print('You are ${use(ageKey)} which is ${use(daysOldKey)} '
+  ///         'count: ${use(countKey)}'));
+  ///   });
   Scope([String? debugName]) {
     _debugName = debugName ?? 'Unnamed Scope - pass debugName to ctor';
   }
@@ -23,9 +46,9 @@ class Scope {
 
   late final String _debugName;
 
-  final provided = <ScopeKey<dynamic>, dynamic>{};
-  final _singles = <ScopeKey<dynamic>, Factory>{};
-  final _sequences = <ScopeKey<dynamic>, Factory>{};
+  final _provided = <ScopeKey<dynamic>, dynamic>{};
+  final _singles = <ScopeKey<dynamic>, _Factory>{};
+  final _sequences = <ScopeKey<dynamic>, _Factory>{};
 
   /// Injects [value] into the [Scope].
   ///
@@ -33,13 +56,15 @@ class Scope {
   /// [use] from anywhere within the action
   /// method provided to [run]
   void value<T>(ScopeKey<T> key, T value) {
-    if (provided.containsKey(key)) {
+    if (_provided.containsKey(key)) {
       throw DuplicateDependencyException(key);
     }
-    provided.putIfAbsent(key, () => value);
+    _provided.putIfAbsent(key, () => value);
   }
 
   @Deprecated('Use single')
+
+  /// Use [single].
   void factory<T>(ScopeKey<T> key, T Function() factory) =>
       single(key, factory);
 
@@ -82,8 +107,8 @@ class Scope {
     _resolveSingles();
 
     return runZoned(action, zoneValues: {
-      Injector:
-          Injector(provided.map<ScopeKey<dynamic>, dynamic>((t, dynamic v) {
+      _Injector:
+          _Injector(_provided.map<ScopeKey<dynamic>, dynamic>((t, dynamic v) {
         if (v is Function) {
           return MapEntry<ScopeKey<dynamic>, dynamic>(t, t._castFunction(v));
         } else {
@@ -94,7 +119,7 @@ class Scope {
   }
 
   void _resolveSingles() {
-    final injector = SingleInjector(_singles);
+    final injector = _SingleInjector(_singles);
     runZoned(() {
       injector.zone = Zone.current;
       // Cause [injector] to call all factories.
@@ -103,7 +128,7 @@ class Scope {
         /// and adding it as a value.
         value<dynamic>(key, injector.get<dynamic>(key));
       }
-    }, zoneValues: {Injector: injector});
+    }, zoneValues: {_Injector: injector});
   }
 
   /// Returns the value provided for [key], or the keys default value if no
@@ -114,16 +139,26 @@ class Scope {
   ///
   /// A [CircularDependencyException] will be thrown if a circular
   /// dependency is discovered values provided by [single] or [sequence].
-  static T use<T>(ScopeKey<T> key) => _use(key);
+  static T use<T>(ScopeKey<T> key, {T Function()? withDefault}) =>
+      _use(key, withDefault: withDefault);
+
+  /// Returns true if [key] is contained within the current [Scope]
+  /// or an ancestor [Scope]
+  ///
+  /// For nullable types even if the value is null [hasScopeKey]
+  /// will return true if a value was injected.
+  static bool hasScopeKey<T>(ScopeKey<T> key) => _hasScopeKey(key);
 
   /// Returns true if [key] is contained within the current scope
-  static bool hasScopeKey<T>(ScopeKey<T> key) => _hasScopeKey(key);
+  /// or an ancestor [Scope] or if the [key] has a default value.
+  ///
+  /// For nullable types even if the value is null [hasScopeValue]
+  /// will return true if a value was injected.
+  static bool hasScopeValue<T>(ScopeKey<T> key) => _hasScopeValue(key);
 
   /// Returns true if the caller is running within a [Scope]
   static bool isWithinScope() => _isWithinScope();
 }
-
-typedef ValueFactory<T> = T? Function();
 
 /// Returns the value provided for [key], or the keys default value if no
 /// value was provided.
@@ -132,13 +167,25 @@ typedef ValueFactory<T> = T? Function();
 /// is not in scope.
 ///
 /// A [CircularDependencyException] will be thrown if a circular
-/// dependency is discovered values provided by [single] or [sequence].
-T use<T>(ScopeKey<T> key) => _use(key);
+/// dependency is discovered values provided by [Scope.single]
+/// or [Scope.sequence].
+T use<T>(ScopeKey<T> key, {T Function()? withDefault}) =>
+    _use(key, withDefault: withDefault);
 
-T _use<T>(ScopeKey<T> key) {
+T _use<T>(ScopeKey<T> key, {T Function()? withDefault}) {
   final injector =
-      (Zone.current[Injector] as Injector?) ?? const Injector.empty();
-  final value = injector.get(key);
+      (Zone.current[_Injector] as _Injector?) ?? const _Injector.empty();
+
+  T value;
+  if (_hasScopeKey(key)) {
+    value = injector.get(key);
+  } else if (withDefault != null) {
+    value = withDefault();
+  } else {
+    /// the key may have a default so lets get it
+    /// or throw a MissingDependencyException
+    value = injector.get(key);
+  }
 
   return value;
 }
@@ -146,16 +193,38 @@ T _use<T>(ScopeKey<T> key) {
 /// Returns true if [T] was declared as a nullable type (e.g. String?)
 bool isNullable<T>() => null is T;
 
-/// Returns true if [key] is contained within the current scope.
+/// Returns true if [key] is contained within the current [Scope]
+/// or an ancestor [Scope]
+///
 /// For nullable types even if the value is null [hasScopeKey]
 /// will return true if a value was injected.
 bool hasScopeKey<T>(ScopeKey<T> key) => _hasScopeKey(key);
 
 /// Returns true if [key] is contained within the current scope
+/// or an ancestor [Scope] or if the [key] has a default value.
+///
+/// For nullable types even if the value is null [hasScopeValue]
+/// will return true if a value was injected.
+bool hasScopeValue<T>(ScopeKey<T> key) => _hasScopeValue(key);
+
+/// Returns true if [key] is contained within the current scope
+bool _hasScopeValue<T>(ScopeKey<T> key) {
+  var _hasScopeKey = true;
+  final injector =
+      (Zone.current[_Injector] as _Injector?) ?? const _Injector.empty();
+  if (injector.hasValue(key)) {
+    _hasScopeKey = true;
+  } else {
+    _hasScopeKey = false;
+  }
+  return _hasScopeKey;
+}
+
+/// Returns true if [key] is contained within the current scope
 bool _hasScopeKey<T>(ScopeKey<T> key) {
   var _hasScopeKey = true;
   final injector =
-      (Zone.current[Injector] as Injector?) ?? const Injector.empty();
+      (Zone.current[_Injector] as _Injector?) ?? const _Injector.empty();
   if (injector.hasKey(key)) {
     // final value = injector.get(key);
     // if (isNullable<T>() && value == null) {
@@ -171,4 +240,4 @@ bool _hasScopeKey<T>(ScopeKey<T> key) {
 /// Returns true if the caller is running within a [Scope]
 bool isWithinScope() => _isWithinScope();
 
-bool _isWithinScope() => Zone.current[Injector] != null;
+bool _isWithinScope() => Zone.current[_Injector] != null;